% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_skills.R
\name{detect_skills}
\alias{detect_skills}
\title{Detect and standardize occupational skills in unstructured text data.}
\usage{
detect_skills(data, id, input, output)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of text data that includes the names of skills you want to detect.}

\item{output}{Output column can be named whatever name you want.}
}
\description{
Explanation here.
}
\examples{

library(tidyverse)
library(identidy)
data(skills_data)

skills_data <- skills_data \%>\% 
  rename(skill_name = skill) \%>\% 
  rowid_to_column("rowid")

classified_skills <- skills_data \%>\%
  detect_skills(rowid, skill_name, recoded_skill)

}
